+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+## [4.10.1] - 2026-02-12
+
 ### Changed
 
 - Change cache logic so initial `id_token` doesn't get discarded.
@@ -1641,7 +1643,8 @@ This release supports rendering for CRs:
 - `AppCatalog`
 - `App`
 
-[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v4.10.0...HEAD
+[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v4.10.1...HEAD
+[4.10.1]: https://github.com/giantswarm/kubectl-gs/compare/v4.10.0...v4.10.1
 [4.10.0]: https://github.com/giantswarm/kubectl-gs/compare/v4.9.1...v4.10.0
 [4.9.1]: https://github.com/giantswarm/kubectl-gs/compare/v4.9.0...v4.9.1
 [4.9.0]: https://github.com/giantswarm/kubectl-gs/compare/v4.8.1...v4.9.0